requirements.txt
transformers>=4.36.2
peft>=0.7.1
accelerate>=0.26.1
torch>=2.0
bitsandbytes
Pillow
pandas
tqdm
nltk
rouge
sentence-transformers
